Conversion in place, todos set
