+++ config/manager/kustomization.yaml
@@ -13,4 +13,4 @@ kind: Kustomization
 images:
 - name: controller
   newName: i566483/api-gateway-controller
-  newTag: 2.0.1
+  newTag: 2.0.2
