+++ config/samples/gateway_v1beta1_apirule-svc-per-rule.yaml
@@ -7,7 +7,7 @@ metadata:
     app: orders-service
     example: orders-service
 spec:
-  host: testHost.com
+  host: perrule.testHost.com
   gateway: kyma-system/kyma-gateway
   rules:
     - path: /.*
