+++ config/samples/gateway_v1beta1_apirule-svc-with-override.yaml
@@ -7,7 +7,7 @@ metadata:
     app: orders-service
     example: orders-service
 spec:
-  host: testHost.com
+  host: override.testHost.com
   service:
     name: orders-service
     port: 80
