+++ internal/processing/helpers.go
@@ -1,56 +1,10 @@
 package processing
 
 import (
-	"fmt"
-
-	"github.com/kyma-incubator/api-gateway/internal/helpers"
-
 	gatewayv1beta1 "github.com/kyma-incubator/api-gateway/api/v1beta1"
 	"github.com/kyma-incubator/api-gateway/internal/builders"
-	rulev1alpha1 "github.com/ory/oathkeeper-maester/api/v1alpha1"
 	k8sMeta "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
-
-func modifyAccessRule(existing, required *rulev1alpha1.Rule) {
-	existing.Spec = required.Spec
-}
-
-func generateAccessRule(api *gatewayv1beta1.APIRule, rule gatewayv1beta1.Rule, accessStrategies []*gatewayv1beta1.Authenticator, additionalLabels map[string]string, defaultDomainName string) *rulev1alpha1.Rule {
-	namePrefix := fmt.Sprintf("%s-", api.ObjectMeta.Name)
-	namespace := api.ObjectMeta.Namespace
-	ownerRef := generateOwnerRef(api)
-
-	arBuilder := builders.AccessRule().
-		GenerateName(namePrefix).
-		Namespace(namespace).
-		Owner(builders.OwnerReference().From(&ownerRef)).
-		Spec(builders.AccessRuleSpec().From(generateAccessRuleSpec(api, rule, accessStrategies, defaultDomainName))).
-		Label(OwnerLabel, fmt.Sprintf("%s.%s", api.ObjectMeta.Name, api.ObjectMeta.Namespace))
-
-	for k, v := range additionalLabels {
-		arBuilder.Label(k, v)
-	}
-
-	return arBuilder.Get()
-}
-
-func generateAccessRuleSpec(api *gatewayv1beta1.APIRule, rule gatewayv1beta1.Rule, accessStrategies []*gatewayv1beta1.Authenticator, defaultDomainName string) *rulev1alpha1.RuleSpec {
-	accessRuleSpec := builders.AccessRuleSpec().
-		Match(builders.Match().
-			URL(fmt.Sprintf("<http|https>://%s<%s>", helpers.GetHostWithDomain(*api.Spec.Host, defaultDomainName), rule.Path)).
-			Methods(rule.Methods)).
-		Authorizer(builders.Authorizer().Handler(builders.Handler().
-			Name("allow"))).
-		Authenticators(builders.Authenticators().From(accessStrategies)).
-		Mutators(builders.Mutators().From(rule.Mutators))
-
-	if api.Spec.Service != nil {
-		return accessRuleSpec.Upstream(builders.Upstream().
-			URL(fmt.Sprintf("http://%s.%s.svc.cluster.local:%d", *api.Spec.Service.Name, api.ObjectMeta.Namespace, int(*api.Spec.Service.Port)))).Get()
-	}
-
-	return accessRuleSpec.Get()
-}
 
 func isSecured(rule gatewayv1beta1.Rule) bool {
 	if len(rule.Mutators) > 0 {
