+++ internal/processing/processing.go
@@ -4,10 +4,8 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/kyma-incubator/api-gateway/internal/helpers"
 	"istio.io/api/networking/v1beta1"
 
-	"github.com/kyma-incubator/api-gateway/internal/builders"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/go-logr/logr"
@@ -91,7 +89,7 @@ func (f *Factory) GetActualState(ctx context.Context, api *gatewayv1beta1.APIRul
 		return nil, err
 	}
 
-	//what to do if len(vsList) > 1?
+	//TODO: what to do if len(vsList) > 1?
 
 	if len(vsList.Items) == 1 {
 		state.virtualService = vsList.Items[0]
@@ -198,52 +196,4 @@ func (f *Factory) applyObjDiff(ctx context.Context, objToPatch *objToPatch) erro
 	}
 
 	return nil
-}
-
-func (f *Factory) updateVirtualService(existing, required *networkingv1beta1.VirtualService) {
-	existing.Spec = *required.Spec.DeepCopy()
-}
-
-func (f *Factory) generateVirtualService(api *gatewayv1beta1.APIRule) *networkingv1beta1.VirtualService {
-	virtualServiceNamePrefix := fmt.Sprintf("%s-", api.ObjectMeta.Name)
-	ownerRef := generateOwnerRef(api)
-
-	vsSpecBuilder := builders.VirtualServiceSpec()
-	vsSpecBuilder.Host(helpers.GetHostWithDomain(*api.Spec.Host, f.defaultDomainName))
-	vsSpecBuilder.Gateway(*api.Spec.Gateway)
-
-	for _, rule := range api.Spec.Rules {
-
-		httpRouteBuilder := builders.HTTPRoute()
-		host, port := f.oathkeeperSvc, f.oathkeeperSvcPort
-
-		if !isSecured(rule) {
-			host = fmt.Sprintf("%s.%s.svc.cluster.local", *api.Spec.Service.Name, api.ObjectMeta.Namespace)
-			port = *api.Spec.Service.Port
-		}
-
-		httpRouteBuilder.Route(builders.RouteDestination().Host(host).Port(port))
-		httpRouteBuilder.Match(builders.MatchRequest().Uri().Regex(rule.Path))
-		httpRouteBuilder.CorsPolicy(builders.CorsPolicy().
-			AllowOrigins(f.corsConfig.AllowOrigins...).
-			AllowMethods(f.corsConfig.AllowMethods...).
-			AllowHeaders(f.corsConfig.AllowHeaders...))
-		httpRouteBuilder.Headers(builders.Headers().
-			SetHostHeader(helpers.GetHostWithDomain(*api.Spec.Host, f.defaultDomainName)))
-		vsSpecBuilder.HTTP(httpRouteBuilder)
-	}
-
-	vsBuilder := builders.VirtualService().
-		GenerateName(virtualServiceNamePrefix).
-		Namespace(api.ObjectMeta.Namespace).
-		Owner(builders.OwnerReference().From(&ownerRef)).
-		Label(OwnerLabel, fmt.Sprintf("%s.%s", api.ObjectMeta.Name, api.ObjectMeta.Namespace))
-
-	for k, v := range f.additionalLabels {
-		vsBuilder.Label(k, v)
-	}
-
-	vsBuilder.Spec(vsSpecBuilder)
-
-	return vsBuilder.Get()
 }
